requirements.txt
asgiref==3.4.1
boto3==1.18.26
botocore==1.21.26
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-filter==2.4.0
django-s3direct==0.4.2
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
